Moves DummyEventManager to central useful location
* Used in tests, worthwhile have single class located in central location
