+++ komodo-utils/src/test/java/org/komodo/test/utils/DummyEventManager.java
@@ -1,0 +1,49 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * See the COPYRIGHT.txt file distributed with this work for information
+ * regarding copyright ownership.  Some portions may be licensed
+ * to Red Hat, Inc. under one or more contributor license agreements.
+ * 
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ * 
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+ * 02110-1301 USA.
+ */
+package org.komodo.test.utils;
+
+import org.komodo.spi.runtime.EventManager;
+import org.komodo.spi.runtime.ExecutionConfigurationEvent;
+import org.komodo.spi.runtime.ExecutionConfigurationListener;
+
+public class DummyEventManager implements EventManager {
+
+    @Override
+    public boolean addListener(ExecutionConfigurationListener listener) {
+        return true;
+    }
+
+    @Override
+    public void permitListeners(boolean enable) {
+        // Do Nothing
+    }
+
+    @Override
+    public void notifyListeners(ExecutionConfigurationEvent event) {
+        // Do Nothing
+    }
+
+    @Override
+    public boolean removeListener(ExecutionConfigurationListener listener) {
+        return true;
+    }
+}
