+++ plugins/komodo-plugin-service/src/test/java/org/komodo/osgi/TestPluginServiceTeiidInstances.java
@@ -31,8 +31,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.komodo.spi.query.TeiidService;
 import org.komodo.spi.runtime.EventManager;
-import org.komodo.spi.runtime.ExecutionConfigurationEvent;
-import org.komodo.spi.runtime.ExecutionConfigurationListener;
 import org.komodo.spi.runtime.HostProvider;
 import org.komodo.spi.runtime.TeiidAdminInfo;
 import org.komodo.spi.runtime.TeiidInstance;
@@ -40,33 +38,11 @@ import org.komodo.spi.runtime.TeiidJdbcInfo;
 import org.komodo.spi.runtime.TeiidParent;
 import org.komodo.spi.runtime.TeiidVdb;
 import org.komodo.spi.runtime.version.TeiidVersion;
+import org.komodo.test.utils.DummyEventManager;
 import org.komodo.test.utils.TestUtilities;
 
 @RunWith(Arquillian.class)
 public class TestPluginServiceTeiidInstances extends AbstractTestPluginService {
-
-    protected class DummyEventManager implements EventManager {
-
-        @Override
-        public boolean addListener(ExecutionConfigurationListener listener) {
-            return true;
-        }
-
-        @Override
-        public void permitListeners(boolean enable) {
-            // Do Nothing
-        }
-
-        @Override
-        public void notifyListeners(ExecutionConfigurationEvent event) {
-            // Do Nothing
-        }
-
-        @Override
-        public boolean removeListener(ExecutionConfigurationListener listener) {
-            return true;
-        }
-    }
 
     @Test
     public void testTeiidInstanceConnection() throws Exception {
